requirements.txt
fastapi
uvicorn
python-dotenv
opencv-python
numpy
pillow
requests
python-multipart
beautifulsoup4